move TrackedReader into separate mod and upgrade doc
