+++ Cargo.toml
@@ -2,7 +2,7 @@
 name = "filetrack"
 version = "0.1.1"
 edition = "2021"
-description = "a library for persistent reading of logs"
+description = "persistent logrotated reading and other useful io things"
 license = "MIT"
 readme = "README.md"
 repository = "https://github.com/HectorHW/filetrack"
